+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * [THORN-436] Removed deprecated lrc attributes from tests. Added a LRC manager class.
+* [SKB-1047] Add beamformer daisy chain health to SpsStation rollup.
 
 ## 12.0.4